+++ app/controllers/status_messages_controller.rb
@@ -12,16 +12,6 @@ class StatusMessagesController < ApplicationController
              :json
 
   layout :bookmarklet_layout, :only => :bookmarklet
-  
-  # Define bookmarklet layout depending on whether
-  # user is in mobile or desktop mode
-  def bookmarklet_layout
-    if request.format == :mobile
-      'application'
-    else
-      'blank'
-    end
-  end
 
   # Called when a user clicks "Mention" on a profile page
   # @param person_id [Integer] The id of the person to be mentioned
@@ -88,6 +78,8 @@ class StatusMessagesController < ApplicationController
     end
   end
 
+  private
+
   def destination_aspect_ids
     if params[:status_message][:public] || params[:status_message][:aspect_ids].first == "all_aspects"
       current_user.aspect_ids
@@ -115,5 +107,15 @@ class StatusMessagesController < ApplicationController
 
   def remove_getting_started
     current_user.disable_getting_started
+  end
+
+  # Define bookmarklet layout depending on whether
+  # user is in mobile or desktop mode
+  def bookmarklet_layout
+    if request.format == :mobile
+      'application'
+    else
+      'blank'
+    end
   end
 end