+++ script/get_config.rb
@@ -3,8 +3,18 @@
 # licensed under the Affero General Public License version 3 or later.  See
 # the COPYRIGHT file.
 
+unless ARGV.length >= 1
+  $stderr.puts "Usage: ./script/get_config.rb var=option | option [...]"
+  Process.exit(1)
+end
+
 require 'rubygems'
 require 'pathname'
+
+require 'active_support/core_ext/class/attribute_accessors'
+require 'active_support/core_ext/object/blank'
+require 'active_support/core_ext/module/delegation'
+require 'active_support/core_ext/module/method_names'
 
 class Rails
   def self.root
@@ -14,38 +24,20 @@ class Rails
   def self.env
     env = 'development'
     env = ENV['RAILS_ENV'] if ENV.has_key?('RAILS_ENV')
-    env = ARGV[1] if ARGV.length == 2
     env.downcase
   end
 end
 
+require Rails.root.join("config/load_config")
 
-if ARGV.length >= 1
-  setting_name = ARGV[0]
-  if Rails.env == 'script_server' # load from the special script_server_config.yml file
-    require 'yaml'
-    script_server_config_file = Rails.root.join('config', 'script_server.yml')
-    begin
-      print YAML.load_file(script_server_config_file)['script_server'][setting_name]
-    rescue
-      $stderr.puts "Setting '#{setting_name}' not found in file #{script_server_config_file}."
-      $stderr.puts "Does that file exist? If not, copy it from #{File.basename(script_server_config_file)}.example in the same directory and run this script again."
-      Process.exit(1)
-    end
-  else                            # load from the general diaspora settings file
-    require 'active_support/core_ext/class/attribute_accessors'
-    require 'active_support/core_ext/object/blank'
-    require 'active_support/core_ext/module/delegation'
-    require 'active_support/core_ext/module/method_names'
-    require Rails.root.join("config/load_config")
-    
-    setting = AppConfig.send(setting_name)
-    setting = setting.get if setting.is_a?(Configuration::Proxy)
-    print setting
+ARGV.each do |arg|
+  var, setting_name = arg.split("=")
+  setting_name = var unless setting_name
+  setting = AppConfig.send(setting_name)
+  setting = setting.get if setting.is_a?(Configuration::Proxy)
+  if var != setting_name
+    puts "#{var}=#{setting}"
+  else
+    puts setting
   end
-else
-  $stderr.puts "Usage: ./script/get_config.rb option [section]"
-  $stderr.puts ""
-  $stderr.puts "section defaults to development"
-  Process.exit(1)
 end