+++ spec/controllers/registrations_controller_spec.rb
@@ -112,9 +112,9 @@ describe RegistrationsController do
         flash[:error].should_not be_blank
       end
 
-      it "re-renders the form" do
+      it "redirects back" do
         get :create, @invalid_params
-        response.should render_template("registrations/new")
+        response.should be_redirect
       end
     end
   end