+++ spec/controllers/status_messages_controller_spec.rb
@@ -34,7 +34,7 @@ describe StatusMessagesController do
       doc.xpath('//head').count.should equal 1
       doc.xpath('//body').count.should equal 1
 
-      save_fixture(html_for('body'), 'empty_bookmarklet') 
+      save_fixture(html_for('body'), 'empty_bookmarklet')
     end
 
     it 'accepts get params' do
@@ -50,13 +50,6 @@ describe StatusMessagesController do
       response.should be_success
 
       save_fixture(html_for('body'), 'prefilled_bookmarklet_dirty')
-    end
-  end
-
-  describe '#new_bookmarklet' do
-    it 'works' do
-      get :new_bookmarklet
-      response.should be_success
     end
   end
 
@@ -91,35 +84,35 @@ describe StatusMessagesController do
       response.status.should == 302
       response.should be_redirect
     end
-    
+
     it 'creates with invalid html' do
       post :create, status_message_hash.merge(:status_message => { :text => "0123456789" * 7000 })
       response.status.should == 302
       response.should be_redirect
     end
-    
+
     it 'creates with valid json' do
       post :create, status_message_hash.merge(:format => 'json')
       response.status.should == 201
     end
-    
+
     it 'creates with invalid json' do
       post :create, status_message_hash.merge(:status_message => { :text => "0123456789" * 7000 }, :format => 'json')
       response.status.should == 403
     end
-    
+
     it 'creates with valid mobile' do
       post :create, status_message_hash.merge(:format => 'mobile')
       response.status.should == 302
       response.should be_redirect
     end
-    
+
     it 'creates with invalid mobile' do
       post :create, status_message_hash.merge(:status_message => { :text => "0123456789" * 7000 }, :format => 'mobile')
       response.status.should == 302
       response.should be_redirect
     end
-    
+
     it 'removes getting started from new users' do
       @controller.should_receive(:remove_getting_started)
       post :create, status_message_hash
@@ -224,17 +217,17 @@ describe StatusMessagesController do
     it 'removes the getting started flag from new users' do
       alice.getting_started = true
       alice.save
-      expect{
-        @controller.remove_getting_started
-      }.to change{
+      expect {
+        @controller.send(:remove_getting_started)
+      }.to change {
         alice.reload.getting_started
       }.from(true).to(false)
     end
 
     it 'does nothing for returning users' do
-      expect{
-        @controller.remove_getting_started
-      }.to_not change{
+      expect {
+        @controller.send(:remove_getting_started)
+      }.to_not change {
         alice.reload.getting_started
       }
     end